+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.0 (unreleased)
 
+* use client side websockets (#10)
 * make server transparent to user (#6)
 * [all issues](https://github.com/felixb/ub0r-streaming/issues?q=milestone%3Av0.2.0)